backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
motor==3.3.2

# Async & Task Queue
celery==5.3.4
aiohttp==3.9.1
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-cors==0.0.6
slowapi==0.1.9
cryptography==41.0.7
PyJWT==2.8.1

# ML & Data Processing
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
numpy==1.24.3
scipy==1.11.4
librosa==0.10.0
librosa-espeak-ng==0.0.2

# Transformers & NLP
transformers==4.35.2
onnxruntime==1.16.3
onnx==1.15.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
mediapipe==0.10.3
retinaface==0.0.13
easyocr==1.7.0
pytesseract==0.3.10

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
webrtcvad==2.1.3

# Data Validation
email-validator==2.1.0

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# PDF & Document Generation
reportlab==4.0.7
pypdf==3.17.1

# S3 / MinIO
boto3==1.29.7
minio==7.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Optional: Enterprise Features
python3-saml==1.16.0
google-auth==2.27.0
